move DNSService to network_/dns.py
